refactor: [EL-4435] move useConversationStartersSync to FSD chat hooks
