+++ src/[fsd]/features/chat/lib/hooks/index.js
@@ -3,3 +3,4 @@ export { useEditConversation } from './useEditConversation.hooks';
 export { useInternalToolsConfig } from './useInternalToolsConfig.hooks';
 export { useAttachmentToolChange } from './useAttachmentToolChange.hooks';
 export { useNewConversationAgentAttachment } from './useNewConversationAgentAttachment.hooks';
+export { useConversationStartersSync } from './useConversationStartersSync.hooks';
